requirements.txt
discord.py

aiomysql

pluginbase
